requirements.txt
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
docopt==0.6.2
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
pipreqs==0.4.11
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
yarg==0.1.9
PyYAML==6.0
